+++ R/compute_dispro.R
@@ -1,23 +1,46 @@
-#' Compute (r)OR
+#' Compute disproportionality
 #'
-#' @description `r lifecycle::badge('stable')` compute_or_abcd() computes
+#' @description `r lifecycle::badge('stable')` Computes
 #' bivariate (reporting) Odds-Ratio and Information Component for a drug-adr pair.
 #'
-#' @details Beware that input should be a data.table. Significance in pharmacovigilance
+#' @details Significance in pharmacovigilance
 #' analysis is only defined if the lower bound of the confidence/credibility
 #'  interval is above 1 (i.e. `low_ci > 1`, or `ic_tail > 0`).
 #' Actually, the function computes an Odds-Ratio,
 #' which is not necessarily a **reporting** Odds-Ratio.
 #'
-#' @returns A data.table, with ROR, IC, and their confidence interval (at `1 - alpha`).
-#' Significance of both (as `signif_or` and `signif_ic`).
+#' @returns A data.table, with ROR, IC, and their
+#' confidence/credibility interval (at `1 - alpha`).
+#' Significance of both (as `signif_or` and `signif_ic`, if `export_raw_values` is TRUE).
 #' @param .data The data.table to compute from.
-#' @param y A character vector, one or more variable to explain.
-#' @param x A character vector, one or more explaining variable.
+#' @param y A character vector, one or more variable to explain (usually an adr).
+#' @param x A character vector, one or more explaining variable (usually a drug).
 #' @param alpha Alpha risk.
 #' @param na_format Character string to fill NA values in ror and ci legends.
 #' @param dig Number of digits for rounding (this argument is passed to `cff`)
+#' @param export_raw_values A logical. Should the raw values be exported?
+#' @param min_n_obs A numeric, compute disproportionality only for pairs
+#' with at least `min_n_obs` cases.
 #' @keywords disproportionality
+#' @returns A data.table with columns
+#' \itemize{
+#' \item `y` and `x`, same as input
+#' \item `n_obs` the number of observed cases
+#' \item `n_exp` the number of expected cases
+#' \item `orl` the formatted Odds-Ratio
+#' \item `or_ci` the formatted confidence interval
+#' \item `ic` the Information Component
+#' \item `ic_tail` the tail probability of the IC
+#' \item `ci_level` the confidence interval level
+#' \item Additional columns, if `export_raw_values` is `TRUE`:
+#' \item `a`, `b`, `c`, `d` the counts in the contingency table
+#' \item `std_er` the standard error of the log(OR)
+#' \item `or` the Odds-Ratio
+#' \item `low_ci` the lower bound of the confidence interval
+#' \item `up_ci` the upper bound of the confidence interval
+#' \item `signif_or` the significance of the Odds-Ratio
+#' \item `signif_ic` the significance of the Information Component
+#' }
 #' @export
 #' @importFrom rlang .data
 #' @importFrom rlang .env
@@ -39,14 +62,14 @@
 #'   )
 #'
 #' demo |>
-#'   compute_or_abcd(
+#'   compute_dispro(
 #'     y = "a_colitis",
 #'     x = "nivolumab"
 #'   )
 #'
 #' # You don't have to use the pipe syntax, if you're not familiar
 #'
-#' compute_or_abcd(
+#' compute_dispro(
 #'     .data = demo,
 #'     y = "a_colitis",
 #'     x = "nivolumab"
@@ -59,7 +82,7 @@
 #'   names(ex_$d_drecno)
 #'
 #' demo |>
-#'   compute_or_abcd(
+#'   compute_dispro(
 #'     y = "a_colitis",
 #'     x = many_drugs
 #'   )
@@ -71,18 +94,38 @@
 #'   names(ex_$a_llt)
 #'
 #' demo |>
-#' compute_or_abcd(
+#' compute_dispro(
 #'   y = many_adrs,
 #'   x = many_drugs
 #' )
+#'
+#' # Export raw values if you want to built plots, or other tables.
+#'
+#' demo |>
+#'   compute_dispro(
+#'     y = "a_colitis",
+#'     x = "nivolumab",
+#'     export_raw_values = TRUE
+#'   )
+#'
+#' # Set a minimum number of observed cases to compute disproportionality
+#'
+#' demo |>
+#'  compute_dispro(
+#'  y = "a_colitis",
+#'  x = "nivolumab",
+#'  min_n_obs = 5
+#'  )
 
-compute_or_abcd <-
+compute_dispro <-
   function(.data,
            y,
            x,
            alpha = .05,
            na_format = "-",
-           dig = 2) {
+           dig = 2,
+           export_raw_values = FALSE,
+           min_n_obs = 0) {
     z_val <- qnorm(1 - alpha / 2)
 
     # data mask
@@ -95,6 +138,15 @@ compute_or_abcd <-
     # letters in the contingency table
     cases <-
       letters[1 : 4]
+
+    var_to_export <-
+      if(export_raw_values){
+        c("y", "x", "n_obs", "n_exp", "or", "or_ci", "ic", "ic_tail", "ci_level",
+          "a", "b", "c", "d", "std_er", "or_raw", "low_ci", "up_ci",
+          "signif_or", "signif_ic")
+      } else {
+        c("y", "x", "n_obs", "n_exp", "or", "or_ci", "ic", "ic_tail", "ci_level")
+      }
 
     # extract counts from source table
     c_or_abcd_core <-
@@ -161,6 +213,7 @@ compute_or_abcd <-
     ad_aa_counts |>
       dplyr::mutate(
         dplyr::across(dplyr::all_of(c("a", "b", "c", "d")), ~ as.numeric(.x)),
+        n_obs = .data$a,
         n_exp = (.data$a + .data$b) * # n drug
                 (.data$a + .data$c) / # n event
                 (.data$a + .data$b + .data$c + .data$d), # n pop
@@ -169,14 +222,14 @@ compute_or_abcd <-
                         (1 / .data$c) +
                         (1 / .data$d)
                       ),
-        or = .data$a * .data$d / (.data$b * .data$c),
-        low_ci = .data$or * exp(- .env$z_val * .data$std_er),
-        up_ci  = .data$or * exp(+ .env$z_val * .data$std_er),
-        orl = ifelse(
-          .data$or %in% c(0, Inf),
+        or_raw = .data$a * .data$d / (.data$b * .data$c),
+        low_ci = .data$or_raw * exp(- .env$z_val * .data$std_er),
+        up_ci  = .data$or_raw * exp(+ .env$z_val * .data$std_er),
+        or = ifelse(
+          .data$or_raw %in% c(0, Inf),
           .env$na_format,
           cff(
-            num = .data$or,
+            num = .data$or_raw,
             dig = .env$dig,
             method = "num_only"
           )
@@ -199,6 +252,29 @@ compute_or_abcd <-
         ),
         ci_level  = paste0((1 - .env$alpha) * 100, "%"),
         signif_or = ifelse(.data$low_ci  > 1, 1, 0),
-        signif_ic = ifelse(.data$ic_tail > 0, 1, 0)
-      )
+        signif_ic = ifelse(.data$ic_tail > 0, 1, 0),
+        # don't show results for pairs with less than min_n_obs
+        dplyr::across(
+          dplyr::all_of(
+            c("n_exp", "ic", "ic_tail",
+              "a", "b", "c", "d", "std_er", "or_raw", "low_ci", "up_ci",
+              "signif_or", "signif_ic")
+          ),
+          function(num_col)
+            dplyr::if_else(.data$a < .env$min_n_obs,
+                    NA_real_,
+                    num_col)
+        ),
+        dplyr::across(
+          dplyr::all_of(
+            c("or", "or_ci", "ci_level"
+              )
+          ),
+          function(chr_col)
+            dplyr::if_else(.data$a < .env$min_n_obs,
+                           .env$na_format,
+                           chr_col)
+        )
+      ) |>
+      dplyr::select(dplyr::all_of(var_to_export))
   }